fix/refactor addCategoryDialog
